requirements.txt
beautifulsoup4==4.13.4
matplotlib==3.7.5
modelscope==1.23.1
numpy==1.24.4
openai==1.86.0
pandas==2.0.3
pytorch_lightning==2.4.0
Requests==2.32.4
scikit_learn==1.3.2
selenium==4.33.0
swanlab==0.6.3
# torch==2.4.1+cu118
tqdm==4.67.1
transformers==4.46.3
# vllm==0.9.1
argparse
modelscope
